refactor: rename "Vertical Scaler" to "Worker Scaler"
This commit renames all references to "Vertical Scaler" to "Worker Scaler"
throughout the codebase to better reflect the functionality:

- Renamed lib/dynamic-workers-scaler.js to lib/worker-scaler.js
- Renamed kLastVerticalScalerELU symbol to kLastWorkerScalerELU
- Renamed test files from vertical-scaler-*.test.js to worker-scaler-*.test.js
- Renamed fixture directories from vertical-scaler to worker-scaler
- Renamed fixture config files from platformatic.vertical-scaler.json to platformatic.worker-scaler.json
- Updated all references in test files and imports
- Updated user-facing warning message in runtime.js

The deprecated `verticalScaler` config property is kept for backward compatibility
and continues to migrate to the new `workers` configuration.
