+++ packages/runtime/fixtures/worker-scaler-service/platformatic.json
