+++ packages/runtime/fixtures/worker-scaler-service/platformatic.worker-scaler.json
