+++ packages/runtime/fixtures/worker-scaler-service/routes/root.cjs
