+++ packages/runtime/fixtures/worker-scaler/platformatic.json
