+++ packages/runtime/fixtures/worker-scaler/platformatic.worker-scaler.json
