+++ packages/runtime/fixtures/worker-scaler/services/service-1/package.json
