+++ packages/runtime/fixtures/worker-scaler/services/service-1/platformatic.json
