+++ packages/runtime/fixtures/worker-scaler/services/service-1/routes/root.cjs
