+++ packages/runtime/fixtures/worker-scaler/services/service-2/platformatic.json
