+++ packages/runtime/index.d.ts
@@ -57,7 +57,7 @@ export module symbols {
   export declare const kHealthCheckTimer: unique symbol
   export declare const kHealthMetricsTimer: unique symbol
   export declare const kLastHealthCheckELU: unique symbol
-  export declare const kLastVerticalScalerELU: unique symbol
+  export declare const kLastWorkerScalerELU: unique symbol
   export declare const kWorkerStatus: unique symbol
   export declare const kWorkerHealthSignals: unique symbol
   export declare const kStderrMarker: string
